python-migration-script/requirements.txt
boto3==1.35.36
pymongo==4.8.0
Pillow==10.4.0
python-dotenv==1.0.1

